By default use preset from Slic3r.ini
The slicing without GUI used to be done using defaults plus ini files
passed in on the command line. With this change the presets selected
in the GUI last are used as the defaults. This plays beter with
Printrun since the GUI is launched to do the "Slicer configuration" and
it is very confusing that after the configuration STLs opened in Printrun
are not sliced using the options that were selected in the GUI.